remake config to use doubles for consistency

+++ src/main/java/dashketch/mods/custom_music_discs/server/ModConfigs.java
@@ -6,7 +6,7 @@ public class ModConfigs {
     private static final ModConfigSpec.Builder BUILDER = new ModConfigSpec.Builder();
 
     public static final ModConfigSpec.BooleanValue JUKEBOX_RANGE_BOOL;
-    public static final ModConfigSpec.IntValue JUKEBOX_RANGE;
+    public static final ModConfigSpec.DoubleValue JUKEBOX_RANGE;
 
     static {
         BUILDER.comment("SERVER SETTINGS");
@@ -14,7 +14,7 @@ public class ModConfigs {
 
         JUKEBOX_RANGE_BOOL = BUILDER.define("jukebox_attentuation_range_is_vanilla", true);
         BUILDER.comment("If above is false");
-        JUKEBOX_RANGE = BUILDER.defineInRange("jukebox_attenuation_range", 64, 16, 128);
+        JUKEBOX_RANGE = BUILDER.defineInRange("jukebox_attenuation_range", 64.0, 16.0, 128.0);
         BUILDER.pop();
     }
